+++ lib/src/ui/pluto_left_frozen_columns.dart
@@ -2,6 +2,8 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:pluto_grid/pluto_grid.dart';
 
+import 'ui.dart';
+
 class PlutoLeftFrozenColumns extends PlutoStatefulWidget {
   final PlutoGridStateManager stateManager;
 
@@ -68,7 +70,7 @@ class PlutoLeftFrozenColumnsState
   }
 
   Widget _buildColumnGroup(PlutoColumnGroupPair e) {
-    return PlutoVisibilityLayoutId(
+    return LayoutId(
       id: e.key,
       child: PlutoBaseColumnGroup(
         stateManager: stateManager,
@@ -81,7 +83,7 @@ class PlutoLeftFrozenColumnsState
   }
 
   Widget _buildColumn(e) {
-    return PlutoVisibilityLayoutId(
+    return LayoutId(
       id: e.field,
       child: PlutoBaseColumn(
         stateManager: stateManager,