+++ lib/src/ui/pluto_left_frozen_rows.dart
@@ -2,6 +2,8 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:pluto_grid/pluto_grid.dart';
 
+import 'ui.dart';
+
 class PlutoLeftFrozenRows extends PlutoStatefulWidget {
   final PlutoGridStateManager stateManager;